pyproject.toml
[tool.poetry]
name = "ml-training"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "ml_training"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
minio = "^7.1.15"
fastapi = "^0.99.1"
rich = "^13.4.2"
pandas = "^2.0.3"
uvicorn = "^0.22.0"
scikit-learn = "^1.3.0"
psycopg2-binary = "^2.9.6"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
